+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed an issue where MultiSafepay orders in state pending_payment could not go to processing if the order was paid through another way.
 
 ## [2.14.1] - 2022-02-22
-
 ### Changed
 - Bump version